# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Move My Car vehicle QR contact API
servers:
  - url: /api
tags:
  - name: health
  - name: auth
  - name: qr-codes
  - name: scans
  - name: stats
  - name: settings
  - name: users
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthStatus" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a dashboard account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "201":
          description: Account created and logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "409": { $ref: "#/components/responses/Error" }
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthMe" }
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /qr-codes:
    get:
      operationId: listQrCodes
      tags: [qr-codes]
      responses:
        "200":
          description: Owned QR codes
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/QrCode" } }
        "401": { $ref: "#/components/responses/Error" }
    post:
      operationId: createQrCode
      tags: [qr-codes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QrCodeInput" }
      responses:
        "201":
          description: Created QR code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrCode" }
  /qr-codes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: number } }
    get:
      operationId: getQrCode
      tags: [qr-codes]
      responses:
        "200":
          description: QR code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrCode" }
    patch:
      operationId: updateQrCode
      tags: [qr-codes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QrCodeUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrCode" }
    delete:
      operationId: deleteQrCode
      tags: [qr-codes]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /scan/{qrId}:
    post:
      operationId: logScan
      tags: [scans]
      parameters:
        - { name: qrId, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: Public contact options
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicScan" }
        "404": { $ref: "#/components/responses/Error" }
  /scan/{qrId}/action:
    post:
      operationId: logScanAction
      tags: [scans]
      parameters:
        - { name: qrId, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScanActionInput" }
      responses:
        "200":
          description: Logged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /scan-logs:
    get:
      operationId: listScanLogs
      tags: [scans]
      parameters:
        - { name: limit, in: query, schema: { type: number, default: 20 } }
        - { name: qrCodeId, in: query, schema: { type: number } }
      responses:
        "200":
          description: Recent logs
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ScanLog" } }
  /stats:
    get:
      operationId: getStats
      tags: [stats]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stats" }
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
    put:
      operationId: updateSettings
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsInput" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: Dashboard accounts
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/User" } }
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties: { status: { type: string } }
    ErrorResponse:
      type: object
      required: [error]
      properties: { error: { type: string } }
    OkResponse:
      type: object
      required: [ok]
      properties: { ok: { type: boolean } }
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3 }
        password: { type: string, minLength: 1 }
    LoginResult:
      type: object
      required: [ok, username]
      properties:
        ok: { type: boolean }
        username: { type: string }
    AuthMe:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
        userId: { type: ["number", "null"] }
        username: { type: ["string", "null"] }
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id: { type: number }
        username: { type: string }
        createdAt: { type: string, format: date-time }
    UserInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3 }
        password: { type: string, minLength: 6 }
    QrCode:
      type: object
      required: [id, label, location, active, createdAt]
      properties:
        id: { type: number }
        label: { type: string }
        location: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        forwardPhone: { type: ["string", "null"] }
    QrCodeInput:
      type: object
      required: [label, location]
      properties:
        label: { type: string, minLength: 1 }
        location: { type: string, minLength: 1 }
        forwardPhone: { type: ["string", "null"] }
    QrCodeUpdate:
      type: object
      properties:
        label: { type: string, minLength: 1 }
        location: { type: string, minLength: 1 }
        active: { type: boolean }
        forwardPhone: { type: ["string", "null"] }
    PublicScan:
      type: object
      required: [qrCodeId, label, location, active, scanLogId]
      properties:
        qrCodeId: { type: number }
        label: { type: string }
        location: { type: string }
        active: { type: boolean }
        scanLogId: { type: number }
        twilioPhone: { type: ["string", "null"] }
        directPhone: { type: ["string", "null"] }
    ScanActionInput:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [call, text] }
        scanLogId: { type: ["number", "null"] }
    ScanLog:
      type: object
      required: [id, qrCodeId, action, scannedAt]
      properties:
        id: { type: number }
        qrCodeId: { type: number }
        action: { type: string, enum: [scan, call, text] }
        scannedAt: { type: string, format: date-time }
        label: { type: string }
    Stats:
      type: object
      required: [totalQrCodes, activeQrCodes, totalScans, totalCalls, totalTexts, recentLogs]
      properties:
        totalQrCodes: { type: number }
        activeQrCodes: { type: number }
        totalScans: { type: number }
        totalCalls: { type: number }
        totalTexts: { type: number }
        recentLogs: { type: array, items: { $ref: "#/components/schemas/ScanLog" } }
    Settings:
      type: object
      required: [ownerPhone]
      properties:
        ownerPhone: { type: ["string", "null"] }
    SettingsInput:
      type: object
      required: [ownerPhone]
      properties:
        ownerPhone: { type: ["string", "null"] }